memtx: make tree iterators stable -- BROKEN DON'T MERGE
Make memtx tree iterator lookup lazy - in the first 'next' call.
It makes the iterator similar to vinyl in case when a new tuple
is inserted between a moment of iterator creation and a moment
of first tuple retrieval. Make the inserted tuple visible to
the iterator.

Store last returned tuple in the iterator and if the tree is
changed between iterations - search for the stored tuple and
continue iteration from the found position. This will make the
iteration stable and every next iteration will return proper
(next or previous) tuple.

Fix #1796

A test case:

> box.cfg()
> box.space._space:insert{512, 0, 'fuck', 'memtx', 2, { omg = 1}}

Thread 1 "tarantool" received signal SIGSEGV, Segmentation fault.
0x00005555557c9b3d in matras_touch (m=0x555555c28810, id=0) at /home/roman/tarantool/1.6/src/lib/small/small/matras.c:426
426             char *extent3p = (char *)extent2p[n2];
(gdb) bt
    at /home/roman/tarantool/1.6/src/lib/salad/bps_tree.h:2461
    inserted_in_block=0x7ffff041f908, inserted_in_pos=0x7ffff041f906) at /home/roman/tarantool/1.6/src/lib/salad/bps_tree.h:3349
    at /home/roman/tarantool/1.6/src/lib/salad/bps_tree.h:4446
    at /home/roman/tarantool/1.6/src/box/memtx_tree.cc:391
    at /home/roman/tarantool/1.6/src/box/memtx_space.cc:218
    at /home/roman/tarantool/1.6/src/box/memtx_space.cc:428
    at /home/roman/tarantool/1.6/src/box/box.cc:183
    at /home/roman/tarantool/1.6/src/box/box.cc:794
    at /home/roman/tarantool/1.6/src/fiber.h:644