alter: rebuild some secondary indexes if the primary was changed
Since the secondary indexes order will be able to depend on
the order in the primary index, we need to rebuild them if the
primary index was changed.

Only non-unique tree indexes are expected to depend on the
primary index, so rebuild only non-unique tree indexes.

Add a test.

Needed for #2476