box: allow to install space.on_commit triggers in Lua
When space.on_replace trigger is called, it's unclear whether the
statement will actually make it to the space or will be rolled back.
Sometimes, one might need a trigger which is only called after a
successful insertion into a space, e.g. it might be useful for
maintaining Lua objects in sync with space contents. This patch adds
such trigger. Note, if the trigger function installed by the user
yields, the behavior is undefined.

Needed for #389