box: persistent sequences and auto increment
This patch implements a new object type, persistent sequences. Sequences
can be standalone and bound to spaces. Standalone sequences are created
with function box.schema.sequence.create(name, opts). Options include
min/max, start value, increment, just like in Postgresql. All sequences
can be accessed via box.sequence.<name>, similarly to spaces.  To
generate a sequence value, use seq:next() method. A sequence value can
also be reset to the start value or to any other value using seq:reset()
and seq:set() methods.

To create a sequence bound to a space, pass 'auto_increment=true' when
creating the primary index for the space. This only works if the primary
index parts are {1, 'unsigned'}. An auto-increment sequence is named
after the space they belong to: '_auto_increment_<space_name>'. It is
deleted when the primary index is dropped. If the space is marked as
auto incremented, one can insert 'nil' for the primary key value, in
which case the bound sequence will be used to generate the value. If a
value inserted into a space turns out to be greater than the current
sequence value, the sequence is updated accordingly.

Issues:
 - Sequences do not work in vinyl transactions, only in memtx.
 - Permission checks haven't been implemented yet.

Closes #389