sql: make indexes covering, fix scan cost, reorder indexes list
The decision to reorder indexes list was made due to several reasons:
 - indexes are sorted this way in Tarantool internal representation
 - if all indexes become covering, than Tarantool should try pk with
     higher priority, but it seems that in some cases sqlite chooses
     the first suitable index from the indexes list

Tarantool indexes should be covering by nature, because it's representation
is just a pointer to a tuple (containing all columns of the row)

This change led to reading data from secondary indexes on scan if possible.
To save sense and increase usage of locality of date it was necessary to
change cost calculation formula.
By now scan over pk is allways better than scan over secondary index for
Tarantool table.

This change also affected some tests. All result changes was reviewed and
none of them was inefficient.
main changes:
 - "COVERING" word appeared in each query plan statement
   (because indexes became covering)
 - sqlite engine changed some decisions on which index to use for full scan
   (this affected to order of results in some queries)

closes: #2110